Eriytä vähäinen menettely -testi omakseen. Päivitä snapit
